complete role permission

+++ app/Http/Controllers/Backend/RolesController.php
@@ -8,10 +8,22 @@ use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use Illuminate\Support\Facades\DB;
 use App\Models\User;
+use App\Models\Admin;
+use Illuminate\Support\Facades\Auth;
 
 
 class RolesController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -19,6 +31,9 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('role.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
         $roles = Role::all();
         return view('Backend.Pages.Roles.index', compact('roles'));
     }
@@ -30,6 +45,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
         $all_permissions = Permission::orderBy('id', 'desc')->get();
         $permission_groups = User::getpermissionGroups();
 
@@ -44,6 +62,9 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
         // ['guard_name', 'web'];
         // $role = new Role();
         // $role->name = $request->name;
@@ -58,14 +79,18 @@ class RolesController extends Controller
             'name.required' => 'Please give Role Name'
         ]);
 
-        $role = Role::create(['name' => $request->name]);
+        $role = Role::create(['name' => $request->name, 'guard_name' => 'admin']);
+        // $role = new Role();
+        // $role->name = $request->name;
+        $role->guard_name = $request->guardName;
+        $role->save();
 
         $permissions = $request->input('permissions');
         if (!empty($permissions)) {
             $role->syncPermissions($permissions);
         }
         // session()->flash('success', 'Role has been created !!');
-       
+
         return redirect()->route('roles.index');
     }
 
@@ -88,7 +113,10 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
+        $role = Role::findById($id, 'admin');
         $permissions = Permission::all();
         $all_permissions = Permission::all();
         $permission_groups = User::getpermissionGroups();
@@ -105,14 +133,17 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
-         // Validation Data
-         $request->validate([
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
+        // Validation Data
+        $request->validate([
             'name' => 'required|max:100|unique:roles,name,' . $id
         ], [
             'name.requried' => 'Please give a role name'
         ]);
 
-        $role = Role::findById($id);
+        $role = Role::findById($id, 'admin');
         $permissions = $request->input('permissions');
 
         if (!empty($permissions)) {
@@ -131,7 +162,10 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
-        $role = Role::findById($id);
+        if (is_null($this->user) || !$this->user->can('role.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
+        $role = Role::findById($id, 'admin');
 
         if (!is_null($role)) {
             $role->delete();
@@ -140,14 +174,17 @@ class RolesController extends Controller
         return redirect()->route('roles.index');
     }
 
-    public function permission(){
-        $role = Role::all();
-        return view('Backend.Pages.Roles.createPermission', compact('role'));
+    public function permission()
+    {
+        $role = Permission::orderBy('id', 'asc')->select('group_name')->distinct()->get();
+        return view('Backend.Pages.Roles.createPermissions', compact('role'));
     }
-    public function permissionStore(Request $request){
+    public function permissionStore(Request $request)
+    {
         $role = new permission();
         $role->name = $request->name;
         $role->group_name = $request->permission;
+        $role->guard_name = $request->guardName;
         $role->save();
         return redirect()->route('permission');
     }
